Minor fix for data type and length

+++ flame/data.py
@@ -395,25 +395,25 @@ class DataCollatorForLanguageModeling:
                 if self.tokenizer.add_bos_token:
                     offsets = []
                     if batch['input_ids'][0, 0] != self.tokenizer.bos_token_id:
-                        offsets.append(torch.tensor([0], dtype=torch.int32))
-                    offsets.append(torch.where(batch['input_ids'].eq(self.tokenizer.bos_token_id))[1].to(torch.int32))
-                    offsets.append(torch.tensor([len(batch['input_ids'][0])], dtype=torch.int32))
-                    batch['offsets'] = torch.cat(offsets, dim=0)
+                        offsets.append(torch.tensor([0]))
+                    offsets.append(torch.where(batch['input_ids'].eq(self.tokenizer.bos_token_id))[1])
+                    offsets.append(torch.tensor([len(batch['input_ids'][0])]))
+                    batch['offsets'] = torch.cat(offsets, dim=0).to(dtype=torch.int32)
                 elif self.tokenizer.add_eos_token:
-                    offsets = [torch.tensor([0], dtype=torch.int32)]
-                    offsets.append(torch.where(batch['input_ids'].eq(self.tokenizer.eos_token_id))[1].to(torch.int32) + 1)
+                    offsets = [torch.tensor([0])]
+                    offsets.append(torch.where(batch['input_ids'].eq(self.tokenizer.eos_token_id))[1] + 1)
                     if batch['input_ids'][0, -1] != self.tokenizer.eos_token_id:
-                        offsets.append(torch.tensor([len(batch['input_ids'][0])], dtype=torch.int32))
-                    batch['offsets'] = torch.cat(offsets, dim=0)
+                        offsets.append(torch.tensor([len(batch['input_ids'][0])]))
+                    batch['offsets'] = torch.cat(offsets, dim=0).to(dtype=torch.int32)
                 else:
                     raise ValueError("You must allow the tokenizer to add either a bos or eos token as separators.")
             if self.context_len is not None:
                 bos = batch['offsets'][:-1].tolist()
                 eos = batch['offsets'][1:].tolist()
                 batch['offsets'] = torch.cat(
-                    [torch.arange(i, j, self.context_len, dtype=torch.int32) for i, j in zip(bos, eos)] +
-                    [torch.tensor([len(batch['input_ids'][0])], dtype=torch.int32)]
-                )
+                    [torch.arange(i, j, self.context_len) for i, j in zip(bos, eos)] +
+                    [torch.tensor([len(batch['input_ids'][0])])]
+                ).to(dtype=torch.int32)
 
         labels = batch['input_ids'].clone()
         if self.tokenizer.pad_token_id is not None:
